refactor get_mission_path import
